requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

python-dotenv>=1.0.1

strands-agents>=1.0.0
strands-agents-tools>=0.2.16
ollama
